Feat : 크롤링 컨트롤러 구현 5

+++ backend/src/main/java/com/ssacretary/db/repository/KeywordRepository.java
@@ -4,7 +4,7 @@ import com.ssacretary.db.entity.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KeywordRepository extends JpaRepository<Keyword,Integer> {
-    Keyword findByKeywordId(Long keywordId);
+    Keyword findByKeywordId(int keywordId);
 
     Keyword findByKeyword(String keyword);
